requirements.txt
TeslaPy==2.4.0
numpy==1.16.2
paho-mqtt==1.6.1

